+++ src/actions/index.ts
@@ -1,10 +1,9 @@
 import { ADD_TODO, TOGGLE_TODO, DELETE_TODO } from'../actions/types/actionTypes'
 
-let nextId = 0;
+
 
 export const addTodo = (title, text) => ({
     type: ADD_TODO,
-    id: nextId++,
     title,
     text,
 })