+++ src/store/index.ts
@@ -1,7 +1,20 @@
-import { createStore } from 'redux';
+import { createStore, applyMiddleware } from 'redux';
 import rootReducer from '../reducers';
+import { persistStore, persistReducer } from 'redux-persist';
+import { createLogger } from 'redux-logger';
+import { AsyncStorage } from 'react-native';
 
+const persistConfig = {
+    key: 'root',
+    storage: AsyncStorage,
+    whitelist: ['todos']
+}
 
+const persistedReducer = persistReducer(persistConfig, rootReducer)
 
-const store = createStore(rootReducer)
-export default store;
+export const store = createStore(
+    persistedReducer, applyMiddleware(createLogger())
+)
+
+export const persistedStore = persistStore(store);
+